Develop

+++ database/migrations/2022_06_16_175155_create_travel_details_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->foreignId('product_id')->constrained();
             $table->foreignId('product_type_id')->constrained();
             $table->integer('weight');
-            $table->integer('total_price');
+            $table->decimal('total_price');
             $table->timestamps();
         });
     }
